MAGE-1051 Group special price test with PricingTest and refactor shared method
