+++ Test/Integration/Product/PricingTest.php
@@ -5,16 +5,15 @@ namespace Algolia\AlgoliaSearch\Test\Integration\Product;
 use Algolia\AlgoliaSearch\Exceptions\AlgoliaException;
 use Algolia\AlgoliaSearch\Exceptions\ExceededRetriesException;
 use Algolia\AlgoliaSearch\Model\Indexer\Product as ProductIndexer;
-use Algolia\AlgoliaSearch\Test\Integration\TestCase;
+use Magento\Catalog\Model\Product;
 use Magento\Framework\Exception\NoSuchEntityException;
 use Magento\Framework\Indexer\IndexerRegistry;
-use function PHPUnit\Framework\assertTrue;
 
 /**
  * @magentoDbIsolation disabled
  * @magentoAppIsolation enabled
  */
-class PricingTest extends TestCase
+class PricingTest extends ProductsIndexingTestCase
 {
 
     /**
@@ -24,6 +23,9 @@ class PricingTest extends TestCase
     protected const PRODUCT_ID_CONFIGURABLE_STANDARD_PRICE = 62;
 
     protected const PRODUCT_ID_CONFIGURABLE_CATALOG_PRICE_RULE = 1903;
+
+    const SPECIAL_PRICE_TEST_PRODUCT_ID = 9;
+
     /**
      * @var array<int, float>
      */
@@ -126,7 +128,7 @@ class PricingTest extends TestCase
     public function testMagentoProductData(int $productId, float $expectedPrice): void
     {
         /**
-         * @var \Magento\Catalog\Model\Product $product
+         * @var Product $product
          */
         $product = $this->objectManager->get('Magento\Catalog\Model\ProductRepository')->getById($productId);
         $this->assertTrue($product->isInStock(), "Product is not in stock");
@@ -144,6 +146,73 @@ class PricingTest extends TestCase
             array_keys(self::ASSERT_PRODUCT_PRICES),
             self::ASSERT_PRODUCT_PRICES
         );
+    }
+
+    public function testSpecialPrice()
+    {
+        $this->productIndexer->execute([self::SPECIAL_PRICE_TEST_PRODUCT_ID]);
+        $this->algoliaHelper->waitLastTask();
+
+        $res = $this->algoliaHelper->getObjects(
+            $this->indexPrefix .
+            'default_products',
+            [(string) self::SPECIAL_PRICE_TEST_PRODUCT_ID]
+        );
+        $algoliaProduct = reset($res['results']);
+
+        if (!$algoliaProduct || !array_key_exists('price', $algoliaProduct)) {
+            $this->markTestIncomplete('Hit was not returned correctly from Algolia. No Hit to run assetions.');
+        }
+
+        $this->assertEquals(32, $algoliaProduct['price']['USD']['default']);
+        $this->assertEquals('', $algoliaProduct['price']['USD']['special_from_date']);
+        $this->assertEquals('', $algoliaProduct['price']['USD']['special_to_date']);
+
+        $specialPrice = 29;
+        $fromDatetime = new \DateTime();
+        $toDatetime = new \DateTime();
+        $priceFrom = $fromDatetime->modify('-2 day')->format('Y-m-d H:i:s');
+        $priceTo = $toDatetime->modify('+2 day')->format('Y-m-d H:i:s');
+
+        $product = $this->objectManager->create(Product::class);
+        $product->load(self::SPECIAL_PRICE_TEST_PRODUCT_ID);
+
+        $product->setCustomAttributes([
+            'special_price' => $specialPrice,
+            'special_from_date' => date($priceFrom),
+            'special_to_date' => date($priceTo),
+        ]);
+        $product->save();
+
+        $this->productIndexer->execute([self::SPECIAL_PRICE_TEST_PRODUCT_ID]);
+        $this->algoliaHelper->waitLastTask();
+
+        $res = $this->algoliaHelper->getObjects(
+            $this->indexPrefix .
+            'default_products',
+            [(string) self::SPECIAL_PRICE_TEST_PRODUCT_ID]
+        );
+        $algoliaProduct = reset($res['results']);
+
+        $this->assertEquals($specialPrice, $algoliaProduct['price']['USD']['default']);
+        $this->assertEquals("$32.00", $algoliaProduct['price']['USD']['default_original_formated']);
+    }
+
+    protected function tearDown(): void
+    {
+        /** @var Product $product */
+        $product = $this->objectManager->create(Product::class);
+        $product->load(self::SPECIAL_PRICE_TEST_PRODUCT_ID);
+
+        $product->setCustomAttributes([
+            'special_price' => null,
+            'special_from_date' => null,
+            'special_to_date' => null,
+        ]);
+        $product->getResource()->saveAttribute($product, 'special_price');
+        $product->save();
+
+        parent::tearDown();
     }
 
 }
